refactor(components): one directory per component group
PageLayout sat loose at the components root while every peer had a directory of
its own, and LogoButton had a directory inside a directory that the home page and
the results frame both reach into.
