+++ src/components/pageLayout/PageLayout.scss
@@ -1,5 +1,5 @@
-@use "../styles/breakpoints" as *;
-@use "../styles/a11y" as *;
+@use "../../styles/breakpoints" as *;
+@use "../../styles/a11y" as *;
 
 .home {
   position: relative;
