+++ src/components/results/ResultsFrame.tsx
@@ -4,9 +4,9 @@ import { useIsWeekDecided } from "../../context/AppDataContext";
 import { PlayerAnalysisContextProvider } from "../../context/PlayerAnalysisContext";
 import { RakMadnessScores } from "../../types/RakMadnessScores";
 import getClasses from "../../utils/getClasses";
-import LogoButton from "../navbar/LogoButton/LogoButton";
+import LogoButton from "../navbar/LogoButton";
 import ScoresNavbar, { ScoresView } from "../navbar/ScoresNavbar";
-import PageLayout from "../PageLayout";
+import PageLayout from "../pageLayout/PageLayout";
 import PlayerAnalysisDialog from "../playerAnalysis/PlayerAnalysisDialog";
 import SkeletonTable from "../table/SkeletonTable";
 import "./ResultsFrame.scss";
